convert sidefig.js to new align-* classes
Note that align-* classes are assigned to the paragraph that contains the images (using knitr_jekyll_setup.R as an external source now) so these alignments are done by searching for img tags located within align-* classes. The imgs are not tagged themselves.

Also, added initial code to prevent overlaps. This could probably use refinement to allow figs to shift up as well as down. Not really anxious to write the spring-based layout code or recursive positioning this will need. (It requires thought, at the very least).